refactor: extract workspace aggregate helpers
